Move test stubs to common folder
