+++ src/common/formatters/__tests__/colorProperties.ts
@@ -1,0 +1,26 @@
+const properties = {
+  colors: {
+    primary: {
+      name: 'primary',
+      value: '#fff',
+      type: 'color',
+    },
+    background: {
+      name: 'background',
+      value: '#F4F4F4FF',
+      type: 'color',
+    },
+    primaryVariant: {
+      name: 'primaryVariant',
+      value: '#000',
+      type: 'color',
+    },
+    secondary: {
+      name: 'secondary',
+      value: '#FF495BFF',
+      type: 'color',
+    },
+  },
+}
+
+export default properties;
