+++ src/common/formatters/__tests__/spacingProperties.ts
@@ -1,0 +1,24 @@
+const properties = {
+  spacings: {
+    xxxs: {
+      name: 'xxxs',
+      value: 4,
+      type: 'spacing',
+      description: 'XXXS',
+    },
+    xxs: {
+      name: 'xxs',
+      value: 8,
+      type: 'spacing',
+      description: 'XXS',
+    },
+    xs: {
+      name: 'xs',
+      value: 12,
+      type: 'spacing',
+      description: 'XS',
+    },
+  },
+}
+
+export default properties;
