+++ src/common/formatters/__tests__/typographyProperties.ts
@@ -1,0 +1,33 @@
+
+const properties = {
+  H1: {
+    value: {
+      fontFamily: 'Roboto',
+      fontWeight: 'Light',
+      lineHeight: 'AUTO',
+      fontSize: 96,
+      letterSpacing: -1.5,
+      paragraphSpacing: 0,
+      textDecoration: 'none',
+      textCase: 'none',
+    },
+    type: 'typography',
+    description: 'Header',
+  },
+  H2: {
+    value: {
+      fontFamily: 'Roboto',
+      fontWeight: 'Light',
+      lineHeight: 'AUTO',
+      fontSize: 60,
+      letterSpacing: -0.5,
+      paragraphSpacing: 0,
+      textDecoration: 'none',
+      textCase: 'none',
+    },
+    type: 'typography',
+    description: 'Header',
+  },
+}
+
+export default properties;
