--- src/ios/formatters/__tests__/constants.ts
@@ -1,1 +1,0 @@
-export const HEADER = '/* Test header */'
